backend/pyproject.toml
[tool.poetry]
name = "vote app"
version = "0.1.0"
description = ""
authors = ["andreistr21 <andstra45@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-stubs = "^4.2.6"
mypy = "^1.6.1"
djangorestframework-stubs = "^3.14.4"
django-cors-headers = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
